// src/generated/models/Tk_hallsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Tk_hallsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Tk_hallsstatecode = keyof typeof Tk_hallsstatecode;
export const Tk_hallsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Tk_hallsstatuscode = keyof typeof Tk_hallsstatuscode;

export interface Tk_hallsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Tk_hallsstatecode;
  statuscode?: Tk_hallsstatuscode;
  timezoneruleversionnumber?: number;
  tk_audiovisual?: boolean;
  tk_disableaccess?: boolean;
  tk_halldescription?: string;
  tk_hallid: string;
  tk_hallimage?: string;
  tk_hallname: string;
  tk_hallref?: string;
  "tk_HotelBranch@odata.bind"?: string;
  tk_smokingroom?: boolean;
  tk_wifi?: boolean;
  utcconversiontimezonecode?: number;
}

export interface Tk_halls extends Tk_hallsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  tk_audiovisualname?: string;
  tk_disableaccessname?: string;
  tk_hallimage_timestamp?: number;
  tk_hallimage_url?: string;
  tk_hallimageid?: string;
  tk_hotelbranchname?: string;
  tk_smokingroomname?: string;
  tk_wifiname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  tk_hotelbranch?: object;
  _tk_hotelbranch_value?: string;
}
export type Tk_hallsUploadColumnName = 'tk_hallimage';